+++ lib/redmine_depending_custom_fields.rb
@@ -2,9 +2,12 @@
 
 require 'redmine'
 require_relative 'redmine_depending_custom_fields/extended_user_format'
+require_relative 'redmine_depending_custom_fields/dependable_list_format'
 
 module RedmineDependingCustomFields
   FIELD_FORMAT_EXTENDED_USER = 'extended_user'
+  FIELD_FORMAT_DEPENDABLE_LIST = 'dependable_list'
+  FIELD_FORMAT_DEPENDABLE_ENUMERATION = 'dependable_enumeration'
 
   def self.register_formats
     formats = Setting.plugin_redmine_depending_custom_fields['enabled_formats'] || []
@@ -16,6 +19,26 @@ module RedmineDependingCustomFields
       end
     else
       Redmine::FieldFormat.delete FIELD_FORMAT_EXTENDED_USER
+    end
+
+    if formats.include?(FIELD_FORMAT_DEPENDABLE_LIST)
+      Redmine::FieldFormat.add FIELD_FORMAT_DEPENDABLE_LIST, DependableListFormat do |format|
+        format.label = :label_dependable_list
+        format.order = 9
+        format.edit_as = 'list'
+      end
+    else
+      Redmine::FieldFormat.delete FIELD_FORMAT_DEPENDABLE_LIST
+    end
+
+    if formats.include?(FIELD_FORMAT_DEPENDABLE_ENUMERATION)
+      Redmine::FieldFormat.add FIELD_FORMAT_DEPENDABLE_ENUMERATION, DependableEnumerationFormat do |format|
+        format.label = :label_dependable_enumeration
+        format.order = 10
+        format.edit_as = 'enumeration'
+      end
+    else
+      Redmine::FieldFormat.delete FIELD_FORMAT_DEPENDABLE_ENUMERATION
     end
   end
 end