+++ CHANGELOG.md
@@ -4,3 +4,10 @@
 
 * Initial release with Extended user custom field format.
 
+## 0.0.2
+
+* Added Dependable list and Dependable key/value list custom field formats.
+* Parent field must be from the same object type.
+  * List fields may depend on regular lists or other dependable lists.
+  * Key/value fields may depend on enumerations or other dependable key/value lists.
+